Control two 7-segment LED displays
5-10-15    This sketch uses an 8-bit shift register, part number SN74HC595 from Texas Instruments.  It allows the control of up to 8 LEDs with just 3 pins on the Arduino.  Building on v2, this v4 exercise now drives TWO 7-segment numeric displays, part number 276-0075 from Radio Shack.  Skipped v3 because it was a special version using bitRead; probably harder to upgrade to 2 shift registers.

Multiple shift registers can be daisy-chained; they act as a single register, releasing all their data together, still with the same 3 Arduino pins.  This v4 sketch tries to implement this. Pin 9 (QH-prime) on the first chip is connected to pin 14 (SER) of the second chip.  The shift and latch pins are connected together.

After chasing a bug for a long time (see comments at clearArray), the program works as intended.  It cycles from 0 to 99.