Streamlined version, all working fine
5-10-15     This sketch uses an 8-bit shift register, part number SN74HC595 from Texas Instruments.  It allows the control of up to 8 LEDs with just 3 pins on the Arduino.
  
This v5 improves the assignment of LED segment data while cycling from 0 to 99.  Uses two nested loops rather than branching for each group of tens.
    
This v5 also uses shorter assignments to define each numeral.
    
Note that this method is hard-coded for a 2-digit display.  Would need a hundreds loop for a 3-digit display.

It may be that this general method can support only up to 4 shift registers.
"Because of it taking in a binary number, it is limited to 32 characters."
There may also be power limits, requiring the use of transistors or ICs
to add more LEDs.  In any case it would be nice to eliminate the resistor
needed for each LED segment; some ICs make this possible.